+++ previous-versions/common/src/sdk/models/operations/listapikeys.ts
@@ -13,6 +13,9 @@ export class ListApiKeysResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     apiKeys?: shared.ApiKeys;
 
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -22,9 +25,15 @@ export class ListApiKeysResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }