+++ previous-versions/common/RELEASES.md
@@ -671,3 +671,13 @@ Based on:
 - [typescript v0.37.0] previous-versions/common
 ### Releases
 - [NPM v0.37.0] https://www.npmjs.com/package/@codat/common/v/0.37.0 - previous-versions/common
+
+## 2023-10-09 01:27:29
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.96.1 (2.150.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.38.0] previous-versions/common
+### Releases
+- [NPM v0.38.0] https://www.npmjs.com/package/@codat/common/v/0.38.0 - previous-versions/common